refactor: extract super tenant district actions
